fix: actually make the injector not make a new app.asar folder and rename app.asar to _app.asar when its already done

+++ scripts/installer.js
@@ -11,23 +11,23 @@ function installFrogcord() {
   const appPath = path.join(resourcesPath, "app.asar");
   const newAppPath = path.join(resourcesPath, "_app.asar");
 
-  exec("taskkill /f /im discord.exe", (error, stdout, stderr) => {
-    if (error) {
-      console.error(`Error: ${error.message}`);
-      return;
-    }
-    if (stderr) {
-      console.error(`Stderr: ${stderr}`);
-      return;
-    }
-    console.log("Please relaunch Discord.");
-  });
+  // Check if both _app.asar file and app.asar folder exist
+  const appAsarExists =
+    fs.existsSync(appPath) && fs.statSync(appPath).isDirectory();
+  const appAsarFileExists =
+    fs.existsSync(newAppPath) && fs.statSync(newAppPath).isFile();
 
-  if (fs.existsSync(appPath)) {
-    fs.renameSync(appPath, newAppPath);
-    fs.mkdirSync(appPath, { recursive: true });
-  } else if (fs.existsSync(newAppPath)) {
-    console.log("Successfully reinjected.");
+  if (appAsarExists && appAsarFileExists) {
+    console.log(
+      "Both _app.asar file and app.asar folder already exist. Skipping renaming and folder creation."
+    );
+  } else {
+    if (fs.existsSync(appPath) && !fs.statSync(appPath).isDirectory()) {
+      fs.renameSync(appPath, newAppPath);
+    }
+    if (!fs.existsSync(appPath)) {
+      fs.mkdirSync(appPath, { recursive: true });
+    }
   }
 
   // Read and write files synchronously
@@ -40,6 +40,18 @@ function installFrogcord() {
     packagePath,
     '{\n  "name": "discord",\n  "main": "index.js"\n}'
   );
+
+  exec("taskkill /f /im discord.exe", (error, stdout, stderr) => {
+    if (error) {
+      console.error(`Error: ${error.message}`);
+      return;
+    }
+    if (stderr) {
+      console.error(`Stderr: ${stderr}`);
+      return;
+    }
+    console.log("Please relaunch Discord.");
+  });
 }
 
 installFrogcord();
